requirements.txt
Django==2.0.4
djangorestframework==3.8.2
numeral==0.1.0.11
